Serialize Snapshot case class instead of SelectedSnapshot
There is no serializer registered for SelectedSnapshot, but Akka-Peristence does register a default Protobuf serializer for Snapshot that delegates the payload serialization to the standard Akka serializers.

Without this change an undocumented hack is required (setting a custom serializer for SelectedSnapshot) to use something else then Java serialization for snapshots.